--- STEP_02/lotto/src/main/java/com/bootcamp/lotto/service/AutoLottoService.java
@@ -1,7 +1,0 @@
-package com.bootcamp.lotto.service;
-
-public class AutoLottoService {
-    /**
-     * 자동 로또 생성 서비스
-     */
-}
