--- STEP_02/lotto/src/main/java/com/bootcamp/lotto/service/NonAutoLottoService.java
@@ -1,7 +1,0 @@
-package com.bootcamp.lotto.service;
-
-public class NonAutoLottoService {
-    /**
-     * 수동 로또 생성 서비스
-     */
-}
